[WIP] [RFC] Simplify the entry point into the workflow runtime.
This drops the weird `WorkflowSession` concept, and the even weirder `onBeforeStart`
function. The new `launchWorkflowIn2` function (to be renamed) just takes a
callback to handle outputs (it's a suspending callback, so we still keep
backpressure) and returns a `StateFlow` of renderings and snapshots. `StateFlow` is
a new Kotlin specialization of `Flow` that always has a value available. This API is
a _lot_ more straightforward and idiomatic. The issue of how to ensure that users
don't miss really fast outputs is solved by requiring an output callback function as
a parameter. The weird issue where, inside of `onBeforeStart`, you had a
`CoroutineScope` receiver but it was different than the one passed into the launch
function is gone – there's now only a single scope, the one you pass in.

This change also rewrites the main workflow loop to be cleaner and (hopefully) more
readable. Notably, in order to synchronously generate the first rendering, the first
render pass is performed synchronously, before even launching the main runtime
coroutine. This means the first render pass may run on a different thread than the
rest, but I don't think anyone cares since most runtimes are started and executed on
the main thread, and `render` functions should all be side-effect-free anyway.

Fixes #463 and fixes #1179.